requirements.txt
streamlit
pandas
numpy
plotly
google-generativeai
joblib
xgboost
imbalanced-learn
requests

# Multi-agent chatbot dependencies
langchain
chromadb
sentence-transformers
